+++ sha1.h
@@ -18,13 +18,13 @@ A million repetitions of "a"
 #ifndef _SHA1
 #define _SHA1
 typedef struct {
-    unsigned long state[5];
-    unsigned long count[2];
+    unsigned int state[5];
+    unsigned int count[2];
     unsigned char buffer[64];
 } SHA1_CTX;
 
 #define DIGEST_SIZE 20
-void SHA1Transform(unsigned long state[5], unsigned char buffer[64]);
+void SHA1Transform(unsigned int state[5], unsigned char buffer[64]);
 void SHA1Init(SHA1_CTX* context);
 void SHA1Update(SHA1_CTX* context, unsigned char* data, unsigned int len);
 void SHA1Final(unsigned char digest[DIGEST_SIZE], SHA1_CTX* context);
